Revert "added the figure numbers to two of  the folders to be consistent with…"
